Jupyter notebook with code for analysis
This notebook produces 3 interactive visualizations based on the rental data provided.

The first visualization is a simple line graph  where one can filter the average rent increase by state using a dropdown menu. The key element is the dropdown feature that allows a user to filter by the state they are interested in.

The second visualization was meant to be used as an alternative to a boxplot to view the distribution of rent increase in each region by year. The key element is the legend that allows the user to filter by region and the tooltip which could be used to hover over the datapoint to reveal city, state, rent price, and percent change.

The final distribution shows the binned results by year for percent change vs dollar change in rent value by year. It can be used to quickly see not only the percent of rent change, but also the dollar value which were both key metrics in the original goals. The key element is being able to filter using the legend to see how the binned values change for each year.